Modified atom screen position to use public integer members (screenX, screenY, screenZ) instead of generating a new Point3f each time a coordinate along one of the dimensions is needed. Eliminates most object generation and garbage collection during repaint operations. Changed z dimension sort to 1) use integers instead of doubles 2) use the built-in Arrays.sort method.

+++ Jmol/src/org/openscience/jmol/Shape.java
@@ -20,6 +20,7 @@
 package org.openscience.jmol;
 
 import java.awt.Graphics;
+import java.util.Comparator;
 
 /**
  * Provides an interface for graphical components so that
@@ -34,6 +35,7 @@ interface Shape {
 
   public void render(Graphics g);
 
-  public double getZ();
+  public int getZ();
 
 }
+
